added user checks to previous contacts and qsl management

+++ application/models/Qsl_model.php
@@ -2,13 +2,13 @@
 class Qsl_model extends CI_Model {
     function getQsoWithQslList() {
         $CI =& get_instance();
-        $CI->load->model('Stations');
-        $station_id = $CI->Stations->find_active();
+        $CI->load->model('logbooks_model');
+        $logbooks_locations_array = $CI->logbooks_model->list_logbook_relationships($this->session->userdata('active_station_logbook'));
 
         $this->db->select('*');
         $this->db->from($this->config->item('table_name'));
         $this->db->join('qsl_images', 'qsl_images.qsoid = ' . $this->config->item('table_name') . '.col_primary_key');
-        $this->db->where('station_id', $station_id);
+        $this->db->where_in('station_id', $logbooks_locations_array);
 
         return $this->db->get();
     }
@@ -16,6 +16,13 @@ class Qsl_model extends CI_Model {
     function getQslForQsoId($id) {
         // Clean ID
         $clean_id = $this->security->xss_clean($id);
+
+        // be sure that QSO belongs to user
+        $CI =& get_instance();
+        $CI->load->model('logbook_model');
+        if (!$CI->logbook_model->check_qso_is_accessible($clean_id)) {
+            return;
+        }
 
         $this->db->select('*');
         $this->db->from('qsl_images');
@@ -25,8 +32,18 @@ class Qsl_model extends CI_Model {
     }
 
     function saveQsl($qsoid, $filename) {
+        // Clean ID
+        $clean_id = $this->security->xss_clean($qsoid);
+
+        // be sure that QSO belongs to user
+        $CI =& get_instance();
+        $CI->load->model('logbook_model');
+        if (!$CI->logbook_model->check_qso_is_accessible($clean_id)) {
+            return;
+        }
+
         $data = array(
-            'qsoid' => $qsoid,
+            'qsoid' => $clean_id,
             'filename' => $filename
         );
 
@@ -39,6 +56,13 @@ class Qsl_model extends CI_Model {
         // Clean ID
         $clean_id = $this->security->xss_clean($id);
 
+        // be sure that QSO belongs to user
+        $CI =& get_instance();
+        $CI->load->model('logbook_model');
+        if (!$CI->logbook_model->check_qso_is_accessible($clean_id)) {
+            return;
+        }
+
         // Delete Mode
         $this->db->delete('qsl_images', array('id' => $clean_id));
     }
@@ -47,6 +71,13 @@ class Qsl_model extends CI_Model {
         // Clean ID
         $clean_id = $this->security->xss_clean($id);
 
+        // be sure that QSO belongs to user
+        $CI =& get_instance();
+        $CI->load->model('logbook_model');
+        if (!$CI->logbook_model->check_qso_is_accessible($clean_id)) {
+            return;
+        }
+
         $this->db->select('filename');
         $this->db->from('qsl_images');
         $this->db->where('id', $clean_id);
@@ -54,14 +85,14 @@ class Qsl_model extends CI_Model {
         return $this->db->get();
     }
 
-	function searchQsos($callsign) {
-		$CI =& get_instance();
-		$CI->load->model('Stations');
-		$station_id = $CI->Stations->find_active();
+    function searchQsos($callsign) {
+        $CI =& get_instance();
+        $CI->load->model('logbooks_model');
+        $logbooks_locations_array = $CI->logbooks_model->list_logbook_relationships($this->session->userdata('active_station_logbook'));
 
 		$this->db->select('*');
 		$this->db->from($this->config->item('table_name'));
-		$this->db->where('station_id', $station_id);
+		$this->db->where_in('station_id', $logbooks_locations_array);
 		$this->db->where('col_call', $callsign);
 
 		return $this->db->get();
@@ -71,8 +102,15 @@ class Qsl_model extends CI_Model {
 		$clean_qsoid = $this->security->xss_clean($qsoid);
 		$clean_filename = $this->security->xss_clean($filename);
 
+		// be sure that QSO belongs to user
+		$CI =& get_instance();
+		$CI->load->model('logbook_model');
+		if (!$CI->logbook_model->check_qso_is_accessible($clean_qsoid)) {
+			return;
+		}
+
 		$data = array(
-			'qsoid' => $qsoid,
+			'qsoid' => $clean_qsoid,
 			'filename' => $filename
 		);
